Add dashboard side menu
