+++ pages/dashboard.tsx
@@ -1,16 +1,16 @@
 import { NextPage } from "next"
 import React, { ReactElement } from "react"
-import { DialogLayout } from "src/layouts"
+import { AuthenticatedLayout, DialogLayout } from "src/layouts"
 import DashboardDialog from "src/components/dashboard/DashboardDialog"
-import Burger from "src/components/burger/Burger"
+import Menu from "src/components/menu/Menu"
 
 const DashboardPage: NextPage = (): ReactElement => {
   return (
     <>
-      <Burger/>
+      <Menu/>
       <DashboardDialog/>
     </>
   )
 }
 
-export default DialogLayout(DashboardPage)
+export default AuthenticatedLayout(DialogLayout(DashboardPage))
